refactor: use client/loading/server lazy load pattern
