+++ src/components/FindWalletProductTable/server.tsx
@@ -2,7 +2,7 @@ import dynamic from "next/dynamic"
 
 import Loading from "./loading"
 
-export default dynamic(() => import("./index"), {
+export default dynamic(() => import("./client"), {
   ssr: false,
   loading: Loading,
 })
